requirements.txt
langchain
openai
pandas
python-dotenv
requests